Source Hyperlinking Working
The hyperlink does come on every page every time a custom inquiry is answered (not for any othe r section). The issue I am facing is that there are some websites (specially older websites) where this is not working. Also, there is no PDF support here (the name of the section will be there but it will not scroll through the PDF when clicked). The following are the websites where I checked:
https://www.apple.com/au/legal/privacy/en-ww/ (works perfectly fine)
https://store.apple.com/Catalog/irl/Images/salespolicies_institution.html (works but always gets dragged to the wrong section)
https://www.deloitte.com/global/en/legal/legal.html (works fine)